+++ executables/geometrytest/main/application.hpp
@@ -47,7 +47,7 @@ public:
 
    void on_key_down     (int key);
    void on_key_up       (int key);
-   void on_mouse_moved  (int x, int y);
+   void on_mouse_moved  (double x, double y);
    void on_mouse_button (int button, int value);
    void on_scroll       (double x, double y);
